帳號系統定義

+++ db/GLT_User/dbo/Tables/Users.sql
@@ -9,10 +9,12 @@
     [DisplayName] NVARCHAR (20) NULL,
     [Depth]       INT           NOT NULL,
     [CreateTime]  DATETIME      CONSTRAINT [DF_Users_CreateTime] DEFAULT (getdate()) NULL,
-    [CreateUser]  BIGINT        CONSTRAINT [DF_Users_CreateUser] DEFAULT ((0)) NULL,
+    [CreateUser]  INT           CONSTRAINT [DF_Users_CreateUser] DEFAULT ((0)) NULL,
     [ModifyTime]  DATETIME      CONSTRAINT [DF_Users_ModifyTime] DEFAULT (getdate()) NULL,
-    [ModifyUser]  BIGINT        CONSTRAINT [DF_Users_ModifyUser] DEFAULT ((0)) NULL,
+    [ModifyUser]  INT           CONSTRAINT [DF_Users_ModifyUser] DEFAULT ((0)) NULL,
     CONSTRAINT [PK_Users] PRIMARY KEY CLUSTERED ([Id] ASC),
     CONSTRAINT [IX_Users] UNIQUE NONCLUSTERED ([CorpId] ASC, [Name] ASC)
 );
+
+
